fix: preserve dtype in mx.where sentinel, scope _compiled to execution context
- Match mx.where sentinel dtype to input (prevents fp16→fp32 promotion)
- Same fix applied to _default_get_logprobs in trainer.py
- Add _compiled kwarg to _extract_grpo_logprobs so each call site
  controls it: _loss_fn passes self._compiled, metrics passes False
- Rewrite tests to inject actual NaN/Inf via model wrappers and verify
  both the raise (default) and sanitization (compiled) behaviors
- Add dtype-preservation test for fp16 models

+++ textpolicy/generation/mlx_generation.py
@@ -441,7 +441,7 @@ def compute_logprobs(
         # in trainer.py for the same pattern.
         selected = mx.where(
             mx.isnan(selected) | mx.isinf(selected),
-            mx.array(-1e6),
+            mx.array(-1e6, dtype=selected.dtype),
             selected,
         )
     else:
